requirements.txt
bitstring==3.1.5
certifi==2016.9.26
click==6.6
future==0.16.0
keyring==8.7
keyrings.alt==1.3
git+https://github.com/ajay-gandhi/pyicloud.git
python-dateutil==2.6.0
pytz==2016.10
requests==2.12.4
six==1.10.0
tzlocal==1.3
